pyproject.toml
[project]
name = "reg01"
version = "2.0.0"
description = "REG CMU AI Assistant — RAG-powered chatbot with Live2D, voice, and admin portal"
requires-python = ">=3.11"
license = { text = "MIT" }

dependencies = [
    # Web Framework
    "fastapi>=0.115,<1",
    "uvicorn[standard]>=0.34,<1",
    "python-socketio>=5.12,<6",
    "python-dotenv>=1.1,<2",
    "python-multipart>=0.0.18,<1",
    "httpx>=0.28,<1",
    "jinja2>=3.1,<4",

    # Media Processing
    "ffmpeg-python>=0.2,<1",
    "pydub>=0.25,<1",
    "speechrecognition>=3.12,<4",

    # TTS
    "edge-tts>=7.0,<8",

    # PDF Processing
    "PyMuPDF>=1.25,<2",
    "pypdf>=5.4,<6",
    "pdfminer.six>=20240706",

    # LLM & AI
    "openai>=1.68,<2",
    "google-genai>=1.5,<2",
    "sentence-transformers>=3.4,<4",
    "torch>=2.6,<3",
    "transformers>=4.49,<5",
    "langdetect>=1.0,<2",
    "tiktoken>=0.8,<1",

    # Vector Database & Search
    "chromadb>=0.6,<1",
    "scikit-learn>=1.6,<2",
    "rank-bm25>=0.2,<1",

    # Thai Language Support
    "pythainlp>=5.0,<6",

    # Database (PostgreSQL + SQLAlchemy 2 ORM)
    "asyncpg>=0.30,<1",
    "sqlalchemy[asyncio]>=2.0,<3",
    "alembic>=1.14,<2",

    # Cache & Queue (Redis)
    "redis>=5.2,<6",

    # Auth (JWT / RBAC)
    "pyjwt[crypto]>=2.10,<3",
    "passlib[bcrypt]>=1.7,<2",

    # Observability (OpenTelemetry)
    "opentelemetry-api>=1.30,<2",
    "opentelemetry-sdk>=1.30,<2",
    "opentelemetry-exporter-otlp>=1.30,<2",
    "opentelemetry-exporter-prometheus>=0.51b0",
    "opentelemetry-instrumentation-fastapi>=0.51b0",
    "opentelemetry-instrumentation-httpx>=0.51b0",
    "opentelemetry-instrumentation-logging>=0.51b0",
    "prometheus-client>=0.21,<1",

    # Utilities
    "tqdm>=4.67,<5",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3,<9",
    "pytest-asyncio>=0.25,<1",
    "pytest-cov>=6.0,<7",
    "ruff>=0.9,<1",
    "mypy>=1.14,<2",
    "pre-commit>=4.1,<5",
]

[project.scripts]
reg01 = "run:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.backends"

# ─── Tool Configuration ─────────────────────────────────────────────────────

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["backend", "app", "router", "memory", "retriever", "queue_manager", "dev"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["backend/dev"]
asyncio_mode = "auto"
